others/training/requirements.txt
facenet-pytorch
mtcnn
numpy
tqdm
Pillow
urllib3
pandas
